Drop 2.4.1 release * Fix issue with `double_tap_timing` and `triple_tap_timing` that weren’t properly decoded.

+++ icm45686/imu/inv_imu_edmp.h
@@ -204,7 +204,7 @@ typedef struct {
 	inv_imu_edmp_tap_num_t  num;
 	inv_imu_edmp_tap_axis_t axis;
 	inv_imu_edmp_tap_dir_t  direction;
-	uint8_t                 double_tap_timing;
+	uint16_t                double_tap_timing;
 } inv_imu_edmp_tap_data_t;
 
 /** @brief Configure EDMP Output Data Rate.
